+++ adapt/src/main/java/io/noties/adapt/ItemViewGroup.java
@@ -1,0 +1,70 @@
+package io.noties.adapt;
+
+import android.view.LayoutInflater;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+
+import java.util.List;
+
+/**
+ * @since 2.3.0-SNAPSHOT
+ */
+public abstract class ItemViewGroup extends Item<ItemViewGroup.Holder>
+        implements HasChildrenItems {
+
+    // todo: common interface to set/get children
+    private List<Item> children;
+
+    protected ItemViewGroup(long id, @NonNull List<Item> children) {
+        super(id);
+        this.children = children;
+    }
+
+    @Override
+    @NonNull
+    public List<Item> getChildren() {
+        return children;
+    }
+
+    @Override
+    public void setChildren(@NonNull List<Item> children) {
+        this.children = children;
+    }
+
+    @NonNull
+    protected abstract ViewGroup createViewGroup(@NonNull LayoutInflater inflater, @NonNull ViewGroup parent);
+
+    @NonNull
+    @Override
+    public Holder createHolder(@NonNull LayoutInflater inflater, @NonNull ViewGroup parent) {
+        return new Holder(createAdaptViewGroup(createViewGroup(inflater, parent)));
+    }
+
+    @Override
+    public void render(@NonNull Holder holder) {
+        final AdaptViewGroup adaptViewGroup = holder.adaptViewGroup;
+        adaptViewGroup.setItems(children);
+        processState(adaptViewGroup);
+    }
+
+    @NonNull
+    protected AdaptViewGroup createAdaptViewGroup(@NonNull ViewGroup group) {
+        return AdaptViewGroup.create(group);
+    }
+
+    protected void processState(@NonNull AdaptViewGroup adaptViewGroup) {
+        // here state of view can be processed, for example if displayed in a RecyclerView
+        // for ex: `ViewState.process(id(), adaptViewGroup.group());`
+    }
+
+    protected static class Holder extends Item.Holder {
+
+        protected final AdaptViewGroup adaptViewGroup;
+
+        protected Holder(@NonNull AdaptViewGroup adaptViewGroup) {
+            super(adaptViewGroup.viewGroup());
+            this.adaptViewGroup = adaptViewGroup;
+        }
+    }
+}